fix_code

+++ src/controllers/userController.js
@@ -11,38 +11,42 @@ const prisma = new PrismaClient();
 
 const userController = () => {
   const createUser = async (req, res, next) => {
-    const { error: validationError } = createUserSchema.validate(req.body);
-    if (validationError) {
-      return res.status(HTTP_STATUS.BAD_REQUEST).json({ error: validationError.details[0].message });
-    }
-  
-    const { email, password } = req.body;
-  
-    try {
-      const existingUser = await prisma.users.findUnique({
-        where: { email }
-      });
-  
-      if (existingUser) {
-        return res.status(HTTP_STATUS.CONFLICT).json({ error: 'Email is already in use' });
+    upload(req, res, async (err) => {
+      if (err) {
+        return res.status(HTTP_STATUS.INTERNAL_SERVER_ERROR).json({ error: 'Error uploading file' });
       }
   
-      const hashedPassword = await encrypt(password);
+      const { error: validationError } = createUserSchema.validate(req.body);
+      if (validationError) {
+        return res.status(HTTP_STATUS.BAD_REQUEST).json({ error: validationError.details[0].message });
+      }
   
-      upload(req, res, async (err) => {
-        if (err) {
-          return next(err);
+      const { email, password } = req.body;
+  
+      try {
+        const existingUser = await prisma.users.findUnique({
+          where: { email },
+        });
+  
+        if (existingUser) {
+          return res.status(HTTP_STATUS.CONFLICT).json({ error: 'Email is already in use' });
         }
-        
+  
+        const hashedPassword = await encrypt(password);
+  
+        if (!req.file || !req.file.location) {
+          return res.status(HTTP_STATUS.INTERNAL_SERVER_ERROR).json({ error: 'File upload failed or file location is undefined' });
+        }
+  
         try {
           const user = await prisma.users.create({
             data: {
               ...req.body,
               password: hashedPassword,
               state: '1',
-              image_Url: req.file.location, 
+              image_Url: req.file.location,
               created_At_dateTime: new Date(),
-              updated_At_dateTime: null
+              updated_At_dateTime: null,
             },
           });
   
@@ -54,13 +58,13 @@ const userController = () => {
         } catch (error) {
           return next(error);
         }
-      });
-    } catch (error) {
-      next(error);
-    } finally {
-      await prisma.$disconnect();
-    }
-  };
+      } catch (error) {
+        return next(error);
+      } finally {
+        await prisma.$disconnect();
+      }
+    });
+  };  
 
   const getUsers = async (_req, res, next) => {
     try {
@@ -100,66 +104,73 @@ const userController = () => {
     if (paramsError) {
       return res.status(HTTP_STATUS.BAD_REQUEST).json({ error: paramsError.details[0].message });
     }
+    
     const { email, password, ...restBody } = req.body;
     const { error: bodyError } = updateUserSchema.validate(restBody);
     if (bodyError) {
       return res.status(HTTP_STATUS.BAD_REQUEST).json({ error: bodyError.details[0].message });
     }
+    
     try {
       const userId = parseInt(req.params.id);
-
+  
       if (req.user.id !== userId) {
         return res.status(HTTP_STATUS.FORBIDDEN).json({ error: 'Access denied. You can only update your own information.' });
       }
-
+  
       const userExist = await prisma.users.findUnique({
-        where: { id: parseInt(req.params.id) }
+        where: { id: userId }
       });
-
+  
       if (!userExist) {
         return res.status(HTTP_STATUS.NOT_FOUND).json({ error: 'User not found' });
       }
-      if(req.file){
-        const deleteKey = userExist.image_Url.split('/').pop();
-        await deleteFile(deleteKey);
-        userData.image_Url = req.file.location
-      }
+  
       const userData = { ...restBody };
-
-      if (password) {
-        userData.password = await encrypt(req.body.password);
+  
+      if (req.file) {
+        if (userExist.image_Url) {
+          const deleteKey = userExist.image_Url.split('/').pop();
+          await deleteFile(deleteKey);
+        }
+  
+        userData.image_Url = req.file.location;
       }
-
+  
+      if (password) {
+        userData.password = await encrypt(password);
+      }
+  
       if (email) {
         const existingUser = await prisma.users.findUnique({
           where: { email }
         });
-
+  
         if (existingUser && existingUser.id !== userId) {
           return res.status(HTTP_STATUS.CONFLICT).json({ error: 'Email is already in use.' });
         }
         userData.email = email;
       }
-        await prisma.users.update({
-          where: { id: userId },
-          data: userData
-        });
-
+  
+      await prisma.users.update({
+        where: { id: userId },
+        data: userData
+      });
+  
       res.status(HTTP_STATUS.NO_CONTENT).send();
     } catch (error) {
       next(error);
-    }
-    finally {
+    } finally {
       await prisma.$disconnect();
     }
-  };
+  };  
 
   const deleteUser = async (req, res, next) => {
     const { error } = idUserSchema.validate(req.params);
     if (error) {
       return res.status(HTTP_STATUS.BAD_REQUEST).json({ error: error.details[0].message });
     }
-
+  
     try {
       const user = await prisma.users.findUnique({
         where: { id: parseInt(req.params.id) }
@@ -167,20 +178,23 @@ const userController = () => {
       if (!user) {
         return res.status(HTTP_STATUS.NOT_FOUND).json({ error: 'User not found' });
       }
+      
+      if (user.image_Url) {
+        const deleteKey = user.image_Url.split('/').pop();
+        await deleteFile(deleteKey);
+      }
+  
       await prisma.users.delete({
         where: { id: parseInt(req.params.id) }
       });
-      const deleteKey = user.image_Url.split('/').pop();
-      await deleteFile(deleteKey);
-
+  
       res.status(HTTP_STATUS.NO_CONTENT).send();
     } catch (error) {
       next(error);
-    }
-    finally {
+    } finally {
       await prisma.$disconnect();
     }
-  };
+  };  
 
   const loginUser = async (req, res, next) => {
     const { email, password } = req.body;
